Extract UncoupledStatement component to his own folder
