+++ react/components/UncoupledStatement/Atoms/ObjectAtom.js
