+++ react/components/UncoupledStatement/Atoms/SubjectAtom.js
