+++ react/components/UncoupledStatement/Atoms/VerbAtom.js
