+++ react/components/UncoupledStatement/UncoupledStatement.js
